[BE] FEAT : 공유사물함 IN_SESSION 상태에서 유저가 대여대기를 취소하는 기능 추가

+++ backend/src/main/java/org/ftclub/cabinet/dto/ShareCodeDto.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 public class ShareCodeDto {
 
-	private final Long shareCode;
+	private final Integer shareCode;
 }
